refactor: refactored outbox query, notifications controllers
