+++ apps/notifications/src/app/controllers/kyc.controller.ts
@@ -1,0 +1,15 @@
+import { Controller } from '@nestjs/common';
+import { EventPattern } from '@nestjs/microservices';
+import { KycRequestDto } from '../dto';
+import { KycService } from '../services/kyc.service';
+import { NOTIFICATIONS_TOPICS } from '@nestjs-booking-clone/common';
+
+@Controller()
+export class KycController {
+  constructor(private readonly kycService: KycService) {}
+
+  @EventPattern(NOTIFICATIONS_TOPICS.KYC.REQUEST_CREATED)
+  async handleKycRequestCreated(data: KycRequestDto) {
+    return this.kycService.handleKycRequestCreated(data);
+  }
+}
